Fixes color scheme changes due to printing.
Chrome will rapidly change color schemes while it renders the page.
Having transitions both cause page flicker, and screws up the state
due to the delay in setting the color scheme. The effect is that doing
a print toggles the color scheme, weirdly enough.

We make a special case for printing where we disable transitions
before and after print rendering. (Complicated by the fact that @media
print is false when restoring.)